Add #29 - 할인 없는 정책의 개념적 혼란을 제거하라
- 기존 공통 할인 정책에선 할인 조건은 반드시 입력되어야 하기에 추상 클래스를 인터페이스로 변경해 할인 금액 계산만 담당하도록 합니다.
	- 공용 = 할인 금액을 계산하세요: Money calculateDiscountAmount(Screening)
- 할인 조건을 입력받는 것은 기본 할인 정책이라는 추상 클래스를 만들어 기존의 할인 정책들을 관리합니다.
	- 기본 할인 정책(DefaultDiscountPolicy)을 만들어라
	- 정률 할인 정책과 비율 할인 정책은 기본 할인 정책을 상속받습니다.
