+++ src/main/java/com/study/object/chapter02/Movie/AmountDiscountPolicy.java
@@ -1,6 +1,6 @@
 package com.study.object.chapter02.Movie;
 
-public class AmountDiscountPolicy extends DiscountPolicy{
+public class AmountDiscountPolicy extends DefaultDiscountPolicy{
     private Money discountAmount;
 
     public AmountDiscountPolicy(Money discountAmount, DiscountCondition... conditions) {
