+++ src/main/java/com/study/object/chapter02/Movie/PercentDiscountPolicy.java
@@ -1,6 +1,6 @@
 package com.study.object.chapter02.Movie;
 
-public class PercentDiscountPolicy extends DiscountPolicy{
+public class PercentDiscountPolicy extends DefaultDiscountPolicy{
     private double percent;
 
     public PercentDiscountPolicy(double percent, DiscountCondition... conditions) {
